move Config to its own file
also move exitCode global into Diagnostics so root.go no longer declares
any globals
